Split the time-consuming integration tests
account -> account + accountUnlock
mempool -> mempool + timelock
onChainBlock -> onChainBlockValid + onChainBlockInvalid
sync -> sync2 + sync3 + sync5
