+++ test/src/integration.long/sync2.test.ts
@@ -20,12 +20,12 @@ import { wait } from "../helper/promise";
 import "mocha";
 import { expect } from "chai";
 
-describe("sync", function() {
+describe("sync 2 nodes", function() {
     const BASE = 600;
-    describe("2 nodes", function() {
-        let nodeA: CodeChain;
-        let nodeB: CodeChain;
+    let nodeA: CodeChain;
+    let nodeB: CodeChain;
 
+    describe("2 nodes", function() {
         beforeEach(async function() {
             nodeA = new CodeChain({ base: BASE });
             nodeB = new CodeChain({ base: BASE });
@@ -296,19 +296,9 @@ describe("sync", function() {
                 });
             });
         });
-
-        afterEach(async function() {
-            if (this.currentTest!.state === "failed") {
-                nodeA.testFailed(this.currentTest!.fullTitle());
-                nodeB.testFailed(this.currentTest!.fullTitle());
-            }
-            await Promise.all([nodeA.clean(), nodeB.clean()]);
-        });
     });
 
-    describe("2 nodes with no parcel relay", function() {
-        let nodeA: CodeChain;
-        let nodeB: CodeChain;
+    describe("with no parcel relay", function() {
         const testSize: number = 5;
 
         beforeEach(async function() {
@@ -342,234 +332,13 @@ describe("sync", function() {
                 (await nodeB.sdk.rpc.chain.getPendingParcels()).length
             ).to.equal(0);
         }).timeout(500 * testSize + 4000);
-
-        afterEach(async function() {
-            if (this.currentTest!.state === "failed") {
-                nodeA.testFailed(this.currentTest!.fullTitle());
-                nodeB.testFailed(this.currentTest!.fullTitle());
-            }
-            await Promise.all([nodeA.clean(), nodeB.clean()]);
-        });
     });
 
-    [3, 5].forEach(function(numNodes) {
-        describe(`${numNodes} nodes`, function() {
-            let nodes: CodeChain[] = [];
-
-            beforeEach(async function() {
-                this.timeout(5000 + 5000 * numNodes);
-
-                for (let i = 0; i < numNodes; i++) {
-                    const node = new CodeChain({
-                        argv: ["--no-discovery"],
-                        base: BASE
-                    });
-                    nodes.push(node);
-                    await node.start();
-                }
-            });
-
-            describe("Connected in a line", function() {
-                describe("All connected", function() {
-                    beforeEach(async function() {
-                        this.timeout(5000 + 5000 * numNodes);
-
-                        const connects = [];
-                        for (let i = 0; i < numNodes - 1; i++) {
-                            connects.push(nodes[i].connect(nodes[i + 1]));
-                        }
-                        await Promise.all(connects);
-                    });
-
-                    it("It should be synced when the first node created a block", async function() {
-                        const parcel = await nodes[0].sendSignedParcel({
-                            awaitInvoice: true
-                        });
-                        for (let i = 1; i < numNodes; i++) {
-                            await nodes[i].waitBlockNumberSync(nodes[i - 1]);
-                            expect(
-                                await nodes[i].getBestBlockHash()
-                            ).to.deep.equal(parcel.blockHash);
-                        }
-                    }).timeout(5000 + 10000 * numNodes);
-
-                    describe("All diverged by both end nodes", function() {
-                        beforeEach(async function() {
-                            const nodeA = nodes[0],
-                                nodeB = nodes[numNodes - 1];
-                            await nodeA.sendSignedParcel();
-                            await nodeB.sendSignedParcel();
-                            expect(await nodeA.getBestBlockNumber()).to.equal(
-                                await nodeB.getBestBlockNumber()
-                            );
-                            expect(
-                                await nodeA.getBestBlockHash()
-                            ).to.not.deep.equal(await nodeB.getBestBlockHash());
-                        });
-
-                        it("Every node should be synced to one", async function() {
-                            const waits = [];
-                            for (let i = 1; i < numNodes; i++) {
-                                waits.push(
-                                    nodes[i].waitBlockNumberSync(nodes[0])
-                                );
-                            }
-                            await Promise.all(waits);
-                        }).timeout(5000 + 5000 * numNodes);
-
-                        it("It should be synced when the first node becomes ahead", async function() {
-                            await nodes[0].sendSignedParcel();
-                            for (let i = 1; i < numNodes; i++) {
-                                await nodes[i].waitBlockNumberSync(
-                                    nodes[i - 1]
-                                );
-                                expect(
-                                    await nodes[i].getBestBlockHash()
-                                ).to.deep.equal(
-                                    await nodes[0].getBestBlockHash()
-                                );
-                            }
-                        }).timeout(5000 + 10000 * numNodes);
-                    });
-                });
-
-                describe("the first node becomes ahead", function() {
-                    beforeEach(async function() {
-                        await nodes[0].sendSignedParcel();
-                    });
-
-                    it("It should be synced when every node connected", async function() {
-                        for (let i = 0; i < numNodes - 1; i++) {
-                            await nodes[i].connect(nodes[i + 1]);
-                            await nodes[i + 1].waitBlockNumberSync(nodes[i]);
-                            expect(
-                                await nodes[i].getBestBlockHash()
-                            ).to.deep.equal(
-                                await nodes[i + 1].getBestBlockHash()
-                            );
-                        }
-                    }).timeout(5000 + 15000 * numNodes);
-                });
-            });
-
-            describe("Connected in a circle", function() {
-                const numHalf: number = Math.floor(numNodes / 2);
-
-                beforeEach(async function() {
-                    this.timeout(5000 + 5000 * numNodes);
-
-                    const connects = [];
-                    for (let i = 0; i < numNodes; i++) {
-                        connects.push(
-                            nodes[i].connect(nodes[(i + 1) % numNodes])
-                        );
-                    }
-                    await Promise.all(connects);
-                });
-
-                it("It should be synced when the first node created a block", async function() {
-                    const parcel = await nodes[0].sendSignedParcel();
-                    for (let i = 1; i <= numHalf; i++) {
-                        await nodes[0].waitBlockNumberSync(nodes[i]);
-                        expect(await nodes[i].getBestBlockHash()).to.deep.equal(
-                            parcel.blockHash
-                        );
-
-                        await nodes[0].waitBlockNumberSync(
-                            nodes[numNodes - i - 1]
-                        );
-                        expect(
-                            await nodes[numNodes - i - 1].getBestBlockHash()
-                        ).to.deep.equal(parcel.blockHash);
-                    }
-                }).timeout(5000 + 5000 * numNodes);
-
-                describe("All diverged by two nodes in the opposite", function() {
-                    beforeEach(async function() {
-                        const nodeA = nodes[0],
-                            nodeB = nodes[numHalf];
-                        await nodeA.sendSignedParcel();
-                        await nodeB.sendSignedParcel();
-                        expect(await nodeA.getBestBlockNumber()).to.equal(
-                            await nodeB.getBestBlockNumber()
-                        );
-                        expect(
-                            await nodeA.getBestBlockHash()
-                        ).to.not.deep.equal(await nodeB.getBestBlockHash());
-                    });
-
-                    it("Every node should be synced", async function() {
-                        const waits = [];
-                        for (let i = 1; i < numNodes; i++) {
-                            waits.push(nodes[i].waitBlockNumberSync(nodes[0]));
-                        }
-                        await Promise.all(waits);
-                    }).timeout(5000 + 5000 * numNodes);
-
-                    it("It should be synced when the first node becomes ahead", async function() {
-                        await nodes[0].sendSignedParcel();
-                        for (let i = 1; i < numNodes; i++) {
-                            await nodes[i].waitBlockNumberSync(nodes[i - 1]);
-                            expect(
-                                await nodes[i].getBestBlockHash()
-                            ).to.deep.equal(await nodes[0].getBestBlockHash());
-                        }
-                    }).timeout(5000 + 10000 * numNodes);
-                });
-            });
-
-            if (numNodes > 3) {
-                describe("Connected in a star", function() {
-                    describe("All connected", function() {
-                        beforeEach(async function() {
-                            this.timeout(5000 + 5000 * numNodes);
-
-                            let connects = [];
-                            for (let i = 1; i < numNodes; i++) {
-                                connects.push(nodes[0].connect(nodes[i]));
-                            }
-                            await Promise.all(connects);
-                        });
-
-                        it("It should be synced when the center node created a block", async function() {
-                            const parcel = await nodes[0].sendSignedParcel();
-                            for (let i = 1; i < numNodes; i++) {
-                                await nodes[0].waitBlockNumberSync(nodes[i]);
-                                expect(
-                                    await nodes[i].getBestBlockHash()
-                                ).to.deep.equal(parcel.blockHash);
-                            }
-                        }).timeout(5000 + 5000 * numNodes);
-
-                        it("It should be synced when one of the outside node created a block", async function() {
-                            const parcel = await nodes[
-                                numNodes - 1
-                            ].sendSignedParcel();
-                            for (let i = 0; i < numNodes - 1; i++) {
-                                await nodes[numNodes - 1].waitBlockNumberSync(
-                                    nodes[i]
-                                );
-                                expect(
-                                    await nodes[i].getBestBlockHash()
-                                ).to.deep.equal(parcel.blockHash);
-                            }
-                        }).timeout(5000 + 10000 * numNodes);
-                    });
-                });
-            }
-
-            afterEach(async function() {
-                this.timeout(5000 + 3000 * numNodes);
-
-                if (this.currentTest!.state === "failed") {
-                    nodes.map(node =>
-                        node.testFailed(this.currentTest!.fullTitle())
-                    );
-                }
-
-                await Promise.all(nodes.map(node => node.clean()));
-                nodes = [];
-            });
-        });
+    afterEach(async function() {
+        if (this.currentTest!.state === "failed") {
+            nodeA.testFailed(this.currentTest!.fullTitle());
+            nodeB.testFailed(this.currentTest!.fullTitle());
+        }
+        await Promise.all([nodeA.clean(), nodeB.clean()]);
     });
 });
